+++ app/src/main/java/me/shikhov/wlogtest/Application.kt
@@ -1,0 +1,11 @@
+package me.shikhov.wlogtest
+
+import android.app.Application
+
+class Application : Application() {
+
+    override fun onCreate() {
+        super.onCreate()
+        System.setProperty("wlog.logLevel", "info")
+    }
+}